move all integration collateral to chipyard
